Move Mistral unit tests for auth to a different file
